Don't use cabal new-style builds.
Revert to v1 cabal commands for now. This is because cabal new-style
builds generate a .ghc.environment file, which can break nix builds.

The file represents all the packages that cabal has figured out it will
use. But these packages differ than the ones specified via nix
(specifically, the package versions in the nix channels). This causes
the build to break, because I suppose nix is running `ghc` or `ghci`,
and these will attempt to load packages specified in the env file.

So the solution is to (1) not `cabal new-build` and (2) delete these
.ghc.environment* files if you do use `new-build`.

Error you may see:

>     Could not load module ‘Distribution.Simple’
>     It is a member of the hidden package ‘Cabal-2.4.0.1’.
>     You can run ‘:set -package Cabal’ to expose it.
>     (Note: this unloads all the modules in the current scope.)
>     Use -v to see a list of the files searched for.
>   |
> 1 | import Distribution.Simple
>   | ^^^^^^^^^^^^^^^^^^^^^^^^^^